Dev 4.0.0

+++ src/ipc/ipc.c
@@ -1,11 +1,14 @@
 #include "ipc.h"
 #include "defines/ipc_values.h"
+#include "defines/settings.h"
+#include "utils/ipUtils.h"
 #include "utils/logger.h"
 #include "utils/memory.h"
 #include "utils/oidc_error.h"
 #include "utils/printer.h"
 #include "utils/stringUtils.h"
 
+#include <arpa/inet.h>
 #include <stdarg.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -16,9 +19,11 @@
 #include <sys/un.h>
 #include <unistd.h>
 
-oidc_error_t initConnectionWithoutPath(struct connection* con, int isServer) {
-  con->server = secAlloc(sizeof(struct sockaddr_un));
-  con->sock   = secAlloc(sizeof(int));
+oidc_error_t initConnectionWithoutPath(struct connection* con, int isServer,
+                                       int tcp) {
+  con->server     = secAlloc(sizeof(struct sockaddr_un));
+  con->tcp_server = secAlloc(sizeof(struct sockaddr_in));
+  con->sock       = secAlloc(sizeof(int));
   if (isServer) {  // msgsock is not needed for a client; furthermore if
                    // the client calls ipc_close it would close stdin
     con->msgsock = secAlloc(sizeof(int));
@@ -26,26 +31,28 @@ oidc_error_t initConnectionWithoutPath(struct connection* con, int isServer) {
   if (con->server == NULL || con->sock == NULL ||
       (con->msgsock == NULL && isServer)) {
     logger(ALERT, "alloc failed\n");
-    exit(EXIT_FAILURE);
+    return oidc_errno;
   }
 
-  *(con->sock) = socket(AF_UNIX, SOCK_STREAM, 0);
+  *(con->sock) = socket(tcp ? AF_INET : AF_UNIX, SOCK_STREAM, 0);
   if (*(con->sock) < 0) {
     logger(ERROR, "opening stream socket: %m");
     oidc_errno = OIDC_ECRSOCK;
     return oidc_errno;
   }
-  con->server->sun_family = AF_UNIX;
+  con->server->sun_family     = AF_UNIX;
+  con->tcp_server->sin_family = AF_INET;
   return OIDC_SUCCESS;
 }
-oidc_error_t initClientConnection(struct connection* con) {
-  return initConnectionWithoutPath(con, 0);
+
+oidc_error_t initClientConnection(struct connection* con, int tcp) {
+  return initConnectionWithoutPath(con, 0, tcp);
 }
 
 oidc_error_t initConnectionWithPath(struct connection* con,
                                     const char*        socket_path) {
   logger(DEBUG, "initializing ipc with path %s\n", socket_path);
-  if (initConnectionWithoutPath(con, 0) != OIDC_SUCCESS) {
+  if (initConnectionWithoutPath(con, 0, 0) != OIDC_SUCCESS) {
     return oidc_errno;
   }
   strcpy(con->server->sun_path, socket_path);
@@ -53,40 +60,64 @@ oidc_error_t initConnectionWithPath(struct connection* con,
 }
 
 /**
- * @brief initializes a client unix domain socket
+ * @brief initializes a client unix domain or tcp socket
  * @param con, a pointer to the connection struct. The relevant fields will be
  * initialized.
  * @param env_var_name, the socket_path environment variable name
  */
-oidc_error_t ipc_client_init(struct connection* con, const char* env_var_name) {
+oidc_error_t ipc_client_init(struct connection* con, unsigned char remote) {
   logger(DEBUG, "initializing client ipc");
-  if (initClientConnection(con) != OIDC_SUCCESS) {
-    return oidc_errno;
-  }
+  const char* env_var_name =
+      remote ? OIDC_REMOTE_SOCK_ENV_NAME : OIDC_SOCK_ENV_NAME;
   const char* path = getenv(env_var_name);
   if (path == NULL) {
     printError("Could not get the socket path from env var '%s'. Have you "
-               "started oidc-agent and set the env var?\n",
+               "set the env var?\n",
                env_var_name);
     logger(WARNING, "Could not get the socket path from env var '%s'",
            env_var_name);
     oidc_errno = OIDC_EENVVAR;
     return OIDC_EENVVAR;
+  }
+
+  if (initClientConnection(con, remote) != OIDC_SUCCESS) {
+    return oidc_errno;
+  }
+
+  if (remote) {
+    logger(DEBUG, "Using TCP socket");
+    char*          tmp_path   = oidc_strcopy(path);
+    char*          ip         = strtok(tmp_path, ":");
+    char*          port_str   = strtok(NULL, ":");
+    unsigned short port       = port_str == NULL ? 0 : strToUShort(port_str);
+    con->tcp_server->sin_port = htons(port ?: 42424);
+    con->tcp_server->sin_addr.s_addr =
+        inet_addr(isValidIP(ip) ? ip : hostnameToIP(ip));
+    secFree(tmp_path);
   } else {
+    logger(DEBUG, "Using UNIX domain socket");
     strcpy(con->server->sun_path, path);
   }
   return OIDC_SUCCESS;
 }
 
 /**
- * @brief connects to a UNIX Domain socket
+ * @brief connects to a UNIX Domain or TCP socket
  * @param con, the connection struct
  * @return the socket or @c OIDC_ECONSOCK on failure
  */
 int ipc_connect(struct connection con) {
-  logger(DEBUG, "connecting ipc %s\n", con.server->sun_path);
-  if (connect(*(con.sock), (struct sockaddr*)con.server,
-              sizeof(struct sockaddr_un)) < 0) {
+  struct sockaddr* server      = (struct sockaddr*)con.server;
+  size_t           server_size = sizeof(struct sockaddr_un);
+  if (con.server->sun_path[0] == '\0') {
+    server      = (struct sockaddr*)con.tcp_server;
+    server_size = sizeof(struct sockaddr_in);
+    logger(DEBUG, "connecting tcp ipc %lu:%hu\n",
+           con.tcp_server->sin_addr.s_addr, con.tcp_server->sin_port);
+  } else {
+    logger(DEBUG, "connecting ipc '%s'\n", con.server->sun_path);
+  }
+  if (connect(*(con.sock), server, server_size) < 0) {
     close(*(con.sock));
     logger(ERROR, "connecting stream socket: %m");
     oidc_errno = OIDC_ECONSOCK;
@@ -249,6 +280,8 @@ oidc_error_t ipc_closeConnection(struct connection* con) {
   }
   secFree(con->server);
   con->server = NULL;
+  secFree(con->tcp_server);
+  con->tcp_server = NULL;
   secFree(con->sock);
   con->sock = NULL;
   secFree(con->msgsock);
@@ -262,8 +295,10 @@ oidc_error_t ipc_closeConnection(struct connection* con) {
  * @return @c OIDC_SUCCESS on success
  */
 oidc_error_t ipc_closeAndUnlinkConnection(struct connection* con) {
-  logger(DEBUG, "Unlinking %s", con->server->sun_path);
-  unlink(con->server->sun_path);
+  if (con->server->sun_path != NULL) {
+    logger(DEBUG, "Unlinking %s", con->server->sun_path);
+    unlink(con->server->sun_path);
+  }
   ipc_closeConnection(con);
   return OIDC_SUCCESS;
 }
